Defer radosgw key provision until OSD's detected
The RADOS gateway on startup will try to initialize a number
of pools in the Ceph cluster; if no OSD's are present at the
point of startup, this operation is re-tried for a period of
5 minutes.

As ceph-radosgw and ceph-mon are typically deployed in LXC
containers, they are up and running before ceph-osd units
are fully operational, resulting in ceph-radosgw units with
no running radosgw process.

By deferring the provision of keys to related ceph-radosgw
units until ceph-osd units are detected, the chance of this
race happening is greatly reduced.

Also add trailing space to ceph.conf as ceph requires that
configuration files end with a newline.

Change-Id: I2a21f021502bde5c688bd3ac4b84ef24a8bdef68
Closes-Bug: 1577519
Closes-Bug: 1578403
(cherry picked from commit fc7c60bf09c38e0114b2e2d00e8f39c51db2cb09)