Add support for user-provided ceph config
Adds a new config-flags option to the charm that
supports setting a dictionary of ceph configuration
settings that will be applied to ceph.conf.

This implementation supports config sections so that
settings can be applied to any section supported by
the ceph.conf template in the charm.

Includes some unit test code cleanup.

Closes-Bug: 1522375
Change-Id: I5fa0890d87425499dbd48af6f2bc1f196713a975