Update tox.ini to stop using unverified package
As of pip 10.0, --allow-unverified is not permitted.

Use of the flag in this repo was previously used to force
installation of python-apt to accommodate certain unit tests.

The unverified package, python-apt, is no longer necessary
for test execution.

Related-Bug: #1760720

Change-Id: Ieb62ac460f605f5599206379b79075704a6fc3e9
(cherry picked from commit 1061abfdc52e4f8596dc253358cd511f964e6319)